+++ src/main/java/org/hestiastore/index/sst/IndexConfigurationDefaultInteger.java
@@ -10,7 +10,7 @@ public class IndexConfigurationDefaultInteger
         implements IndexConfigurationContract {
 
     @Override
-    public long getMaxNumberOfKeysInSegmentCache() {
+    public int getMaxNumberOfKeysInSegmentCache() {
         return 500_000;
     }
